Move summary dashboard and rename about to home
Also adds a warning about PDF issue in Safari
