--- src/containers/About/index.jsx
@@ -1,14 +1,0 @@
-// @flow
-import React from 'react';
-import GeoStreamingHome from '@geostreams/geostreaming/src/containers/Home';
-
-import Partners from '../Partners';
-
-const About = () => (
-    <>
-        <GeoStreamingHome />
-        <Partners />
-    </>
-);
-
-export default About;
